TSE-Bonus-useContext: Add the useContext inside orderContext as a hook +  set the hook inside components
