+++ src/Components/pages/login/LoginForm.jsx
@@ -10,6 +10,7 @@ import { authenticateUser } from "../../../api/user";
 import Welcome from "./Welcome";
 export default function LoginForm() {
   const [username, setUsername] = useState("");
+
   const navigate = useNavigate();
 
   const handleSubmit = async (event) => {
