+++ src/Components/pages/order/Main/Basket/BasketHeader/BasketHeader.jsx
@@ -1,6 +1,5 @@
-import { useContext } from "react";
 import styled from "styled-components";
-import OrderContext from "../../../../../../context/OrderContext";
+import { useOrderContext } from "../../../../../../context/OrderContext";
 import { theme } from "../../../../../../theme/index";
 import { formatPrice } from "../../../../../utils/maths";
 import Header from "../../../../../reusable-ui/Header";
@@ -8,7 +7,7 @@ import { calculateSumToPay } from "./helper";
 import CasinoEffect from "../../../../../reusable-ui/CasinoEffect";
 
 export default function BasketHeader() {
-  const { basket, menu } = useContext(OrderContext);
+  const { basket, menu } = useOrderContext();
 
   const sumToPay = calculateSumToPay(basket, menu);
