+++ src/Components/pages/order/Main/MainRightSide/Admin/Admin.jsx
@@ -1,13 +1,12 @@
 import styled from "styled-components";
-import OrderContext from "../../../../../../context/OrderContext";
+import { useOrderContext } from "../../../../../../context/OrderContext";
 import AdminPanel from "./AdminPanel/AdminPanel";
 import AdminTabs from "./AdminTabs";
-import { useContext } from "react";
 import { fadeInFromBottom } from "../../../../../../theme/animations";
 import { theme } from "../../../../../../theme";
 
 export default function Admin() {
-  const { isCollapsed } = useContext(OrderContext);
+  const { isCollapsed } = useOrderContext();
 
   return (
     <AdminStyled>
