+++ src/Components/pages/order/Main/MainRightSide/Admin/AdminPanel/EditForm.jsx
@@ -1,19 +1,16 @@
-import { useContext, useState } from "react";
-import OrderContext from "../../../../../../../context/OrderContext";
+import { useState } from "react";
+import { useOrderContext } from "../../../../../../../context/OrderContext";
 import Form from "./Form";
 import EditInfoMessage from "./EditInfoMessage";
 import SavingMessage from "../../Menu/SavingMessage";
 import { useSuccessMessage } from "../../../../../../../hooks/useSuccessMessage";
+import { useParams } from "react-router-dom";
 
 export default function EditForm() {
   // state
-  const {
-    username,
-    productSelected,
-    setProductSelected,
-    handleEdit,
-    titleEditRef,
-  } = useContext(OrderContext);
+  const { username } = useParams();
+  const { productSelected, setProductSelected, handleEdit, titleEditRef } =
+    useOrderContext();
 
   const [valueOnFocus, setValueOnFocus] = useState();
   const { isSubmitted: isSaved, displaySuccessMessage } = useSuccessMessage();
