+++ src/Components/pages/order/Main/MainRightSide/MainRightSide.jsx
@@ -1,12 +1,11 @@
-import { useContext } from "react";
 import styled from "styled-components";
-import OrderContext from "../../../../../context/OrderContext";
+import { useOrderContext } from "../../../../../context/OrderContext";
 import { theme } from "../../../../../theme";
 import Admin from "./Admin/Admin";
 import Menu from "./Menu/Menu";
 
 export default function MainRightSide() {
-  const { isModeAdmin } = useContext(OrderContext);
+  const { isModeAdmin } = useOrderContext();
 
   return (
     <MainRightSideStyled>
