+++ src/Components/pages/order/Main/MainRightSide/Menu/Menu.jsx
@@ -1,6 +1,5 @@
-import { useContext } from "react";
 import styled from "styled-components";
-import OrderContext from "../../../../../../context/OrderContext";
+import { useOrderContext } from "../../../../../../context/OrderContext";
 import { theme } from "../../../../../../theme";
 import { formatPrice } from "../../../../../utils/maths";
 import Card from "../../../../../reusable-ui/Card";
@@ -15,10 +14,12 @@ import { isEmpty } from "../../../../../utils/array";
 import Loader from "./Loader";
 import { CSSTransition, TransitionGroup } from "react-transition-group";
 import { menuAnimation } from "../../../../../../theme/animations";
+import { useParams } from "react-router-dom";
 
 export default function Menu() {
+  const { username } = useParams();
+
   const {
-    username,
     menu,
     isModeAdmin,
     handleDelete,
@@ -28,8 +29,7 @@ export default function Menu() {
     handleAddToBasket,
     handleDeleteBasketProduct,
     handleProductSelected,
-  } = useContext(OrderContext);
-  // state
+  } = useOrderContext();
 
   // comportements (gestionnaires d'événement ou "event handlers")
   const handleCardDelete = (event, idProductToDelete) => {
