+++ src/Components/pages/order/OrderPage.jsx
@@ -1,76 +1,32 @@
-import { useEffect, useRef, useState } from "react";
+import { useEffect } from "react";
 import styled from "styled-components";
 import { theme } from "../../../theme";
 import Main from "./Main/Main";
 import NavBar from "./Navbar/NavBar";
-import OrderContext from "../../../context/OrderContext";
-import { EMPTY_PRODUCT } from "../../../enums/product";
-import { useMenu } from "../../../hooks/useMenu";
-import { useBasket } from "../../../hooks/useBasket";
-import { findObjectById } from "../../utils/array";
+import { useOrderContext } from "../../../context/OrderContext";
 import { useParams } from "react-router-dom";
 import { initialiseUserSession } from "./helpers/initialiseUserSession";
 
 export default function OrderPage() {
   // state
-  const [isModeAdmin, setIsModeAdmin] = useState(false);
-  const [isCollapsed, setIsCollapsed] = useState(false);
-  const [currentTabSelected, setCurrentTabSelected] = useState("add");
-  const [newProduct, setNewProduct] = useState(EMPTY_PRODUCT);
-  const [productSelected, setProductSelected] = useState(EMPTY_PRODUCT);
-  const titleEditRef = useRef();
-  const { menu, setMenu, handleAdd, handleDelete, handleEdit, resetMenu } =
-    useMenu();
-  const { basket, setBasket, handleAddToBasket, handleDeleteBasketProduct } =
-    useBasket();
-  const { username } = useParams();
 
-  const handleProductSelected = async (idProductClicked) => {
-    const productClickedOn = findObjectById(idProductClicked, menu);
-    await setIsCollapsed(false);
-    await setCurrentTabSelected("edit");
-    await setProductSelected(productClickedOn);
-    titleEditRef.current.focus();
-  };
+  const { username } = useParams();
+  const { setMenu, setBasket } = useOrderContext();
 
   useEffect(() => {
     initialiseUserSession(username, setMenu, setBasket);
   }, []);
 
-  const orderContextValue = {
-    username,
-    isModeAdmin,
-    setIsModeAdmin,
-    isCollapsed,
-    setIsCollapsed,
-    currentTabSelected,
-    setCurrentTabSelected,
-    menu,
-    handleAdd,
-    handleDelete,
-    resetMenu,
-    newProduct,
-    setNewProduct,
-    productSelected,
-    setProductSelected,
-    handleEdit,
-    titleEditRef,
-    basket,
-    handleAddToBasket,
-    handleDeleteBasketProduct,
-    handleProductSelected,
-  };
+  console.log(username);
 
   //affichage (render)
   return (
-    <OrderContext.Provider value={orderContextValue}>
-      <OrderPageStyled>
-        <div className="container">
-          <NavBar />
-          <Main />
-        </div>
-      </OrderPageStyled>
-    </OrderContext.Provider>
+    <OrderPageStyled>
+      <div className="container">
+        <NavBar />
+        <Main />
+      </div>
+    </OrderPageStyled>
   );
 }
