+++ src/context/OrderContext.jsx
@@ -1,7 +1,11 @@
-import { createContext } from "react";
+import { createContext, useContext, useState, useRef } from "react";
+import { useMenu } from "../hooks/useMenu";
+import { useBasket } from "../hooks/useBasket";
+import { findObjectById } from "../Components/utils/array";
+import { EMPTY_PRODUCT } from "../enums/product";
 
-export default createContext({
-  username: "",
+// 1. création du context
+const OrderContext = createContext({
   isModeAdmin: false,
   setIsModeAdmin: () => {},
 
@@ -30,3 +34,59 @@ export default createContext({
   handleAddToBasket: () => {},
   handleDeleteBasketProduct: () => {},
 });
+
+// 2. création du provider / Initialisation du context
+
+export const OrderContextProvider = ({ children }) => {
+  const [isModeAdmin, setIsModeAdmin] = useState(false);
+  const [isCollapsed, setIsCollapsed] = useState(false);
+  const [currentTabSelected, setCurrentTabSelected] = useState("add");
+  const [newProduct, setNewProduct] = useState(EMPTY_PRODUCT);
+  const [productSelected, setProductSelected] = useState(EMPTY_PRODUCT);
+  const titleEditRef = useRef();
+  const { menu, setMenu, handleAdd, handleDelete, handleEdit, resetMenu } =
+    useMenu();
+  const { basket, setBasket, handleAddToBasket, handleDeleteBasketProduct } =
+    useBasket();
+
+  const handleProductSelected = async (idProductClicked) => {
+    const productClickedOn = findObjectById(idProductClicked, menu);
+    await setIsCollapsed(false);
+    await setCurrentTabSelected("edit");
+    await setProductSelected(productClickedOn);
+    titleEditRef.current.focus();
+  };
+
+  const orderContextValue = {
+    isModeAdmin,
+    setIsModeAdmin,
+    isCollapsed,
+    setIsCollapsed,
+    currentTabSelected,
+    setCurrentTabSelected,
+    menu,
+    setMenu,
+    handleAdd,
+    handleDelete,
+    resetMenu,
+    newProduct,
+    setNewProduct,
+    productSelected,
+    setProductSelected,
+    handleEdit,
+    titleEditRef,
+    basket,
+    setBasket,
+    handleAddToBasket,
+    handleDeleteBasketProduct,
+    handleProductSelected,
+  };
+  return (
+    <OrderContext.Provider value={orderContextValue}>
+      {children}
+    </OrderContext.Provider>
+  );
+};
+
+// 3. création du hook
+export const useOrderContext = () => useContext(OrderContext);
