+++ src/main.jsx
@@ -2,9 +2,12 @@ import ReactDOM from "react-dom/client";
 import App from "./App";
 import "./index.css";
 import { BrowserRouter } from "react-router-dom";
+import { OrderContextProvider } from "./context/OrderContext";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
-  <BrowserRouter>
-    <App />
-  </BrowserRouter>,
+  <OrderContextProvider>
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
+  </OrderContextProvider>,
 );
